Add files via upload
The Old as400 calls was the original that authenticated with POST data.  This version of as400 now has commented code to log into the sales portal, store the sales portal username and session id in the session and the logout function now log users out of the D24 and sales portal webservices.